Mes 3664 pass debrief restructure folders for cat be (#834)
* Mes-3664-restructure-folders-for-pass-debrief-page-cat-be
