+++ src/environment/test-schema-version.ts
@@ -1,0 +1,1 @@
+export const version = '3.0.0';
