+++ src/pages/non-pass-finalisation/non-pass-finalisation.module.ts
@@ -1,8 +1,9 @@
 import { NgModule } from '@angular/core';
 import { IonicPageModule } from 'ionic-angular';
 import { NonPassFinalisationPage } from './non-pass-finalisation';
-import { ComponentsModule } from '../../../components/common/common-components.module';
-import { TestFinalisationComponentsModule } from '../components/test-finalisation.module';
+import { ComponentsModule } from '../../components/common/common-components.module';
+import { TestFinalisationComponentsModule } from
+'../../components/test-finalisation/test-finalisation-component.module';
 
 @NgModule({
   declarations: [
