+++ src/pages/pass-finalisation/pass-finalisation.module.ts
@@ -1,0 +1,10 @@
+import { NgModule } from '@angular/core';
+import { PassFinalisationCatBPageModule } from './cat-b/pass-finalisation.cat-b.page.module';
+
+@NgModule({
+  imports: [
+    PassFinalisationCatBPageModule,
+  ],
+})
+
+export class PassFinalisationPageModule { }
